refactor(alvoc): expose util package functions in public api
